+++ src/spdl/pipeline/defs/_defs.py
@@ -21,7 +21,6 @@ from fractions import Fraction
 from functools import partial
 from typing import Any, Generic, Protocol, runtime_checkable, TypeAlias, TypeVar
 
-from spdl.pipeline._common._convert import _is_isolating_pool
 from spdl.pipeline._common._source_locator import locate_source
 from spdl.pipeline._common._types import _TCallables, _TMergeOp
 
@@ -219,14 +218,8 @@ class PipeConfig(Generic[T, U]):
     def __post_init__(self) -> None:
         op = self._args.op
         if inspect.iscoroutinefunction(op) or inspect.isasyncgenfunction(op):
-            executor = self._args.executor
-            if executor is not None and not _is_isolating_pool(executor):
-                raise ValueError(
-                    "An async op may only be given an isolating-pool (process or interpreter) "
-                    "executor, which is used solely to group the op into a subprocess fusion "
-                    "run -- not to execute it. A non-isolating executor (e.g. a thread pool) "
-                    "has no effect on an async op and is not allowed."
-                )
+            if self._args.executor is not None:
+                raise ValueError("`executor` cannot be specified when op is async.")
         if inspect.isasyncgenfunction(op):
             if self._type == _PipeType.OrderedPipe:
                 raise ValueError(
@@ -427,12 +420,13 @@ class DisaggregateConfig(Generic[T]):
 
 @dataclass(frozen=True)
 class _SubprocessPipelineConfig:
-    """Internal config for a fused run of pipe stages run in a subprocess worker pool.
+    """Internal config for a fused region of pipe stages run in a worker pool.
 
-    Not user-facing. Produced by the fusion pass (``_fuse_subprocess_stages``), which replaces a
-    span of consecutive pool-pipe stages with a single stage that executes the run as one nested
-    pipeline inside the worker pool — eliminating the inter-stage IPC of running each stage on a
-    process/interpreter pool separately.
+    Not user-facing. Produced by the fusion pass (``_fuse_marked_regions``), which replaces a
+    ``.to()`` region — the span of stages between two
+    :py:class:`~spdl.pipeline.defs.PlacementConfig` markers — with a single stage that executes
+    the span as one nested pipeline inside a subprocess (or subinterpreter) worker pool,
+    eliminating the inter-stage IPC of round-tripping data back to the main process.
     """
 
     name: str
@@ -908,11 +902,7 @@ def Pipe(
         executor: A custom executor object to be used to convert the synchronous operation
             into asynchronous one. If ``None``, the default executor is used.
 
-            When ``op`` is already async, ``executor`` must be an isolating-pool (process or
-            interpreter) executor and is not used to run the op (an async op always runs on the
-            event loop) -- it serves only as a subprocess fusion-group tag (see
-            ``fuse_subprocess_stages`` on :py:meth:`~spdl.pipeline.PipelineBuilder.build`).
-            Passing a non-isolating executor (e.g. a thread pool) with an async op is an error.
+            It is invalid to provide this argument when the given op is already async.
         name: The name (prefix) to give to the task.
         output_order: If ``"completion"`` (default), the items are put to output queue
             in the order their process is completed.
